+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Bugfix
 - PSR-0 via composer broken - @Progi1984 GH-51
 - PowerPoint2007 Writer : Fill don't work for RichTextShapes - @Progi1984 GH-61
+- PowerPoint2007 Writer : Hyperlink in table doesn't work - @Progi1984 GH-70
+- ODPresentation Writer : Hyperlink in table doesn't work - @Progi1984 GH-70
 
 ### Miscellaneous